+++ packages/components-angular/src/app/examples/textarea-example-form.component.ts
@@ -1,34 +1,18 @@
 import { ChangeDetectionStrategy, Component } from '@angular/core';
-import { FormControl, FormGroup, FormsModule, ReactiveFormsModule, Validators } from '@angular/forms';
+import { FormControl, FormGroup, FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { PorscheDesignSystemModule } from '@porsche-design-system/components-angular';
 
 @Component({
   selector: 'page-textarea-example-form',
   template: `
-    <form [formGroup]="form" (ngSubmit)="onSubmit()">
+    <form [formGroup]="form" (ngSubmit)="onSubmit()" class="flex flex-col gap-fluid-sm">
       <p-textarea formControlName="myTextarea" [label]="'Some Label'" [required]="true" />
-
-      <button type="submit">Submit</button>
+      <div class="flex gap-fluid-sm">
+        <p-button type="submit">Submit</p-button>
+        <p-button type="reset">Reset</p-button>
+      </div>
+      <p-text>Last submitted data: {{ submittedValue }}</p-text>
     </form>
-
-    <button type="button" (click)="setValue()">
-      Set Value
-    </button>
-
-    <button type="button" (click)="resetValue()">
-      Reset
-    </button>
-
-    <button type="button" (click)="toggleDisabled()">
-      {{ form.controls.myTextarea.disabled ? 'Enable' : 'Disable' }}
-    </button>
-
-    <div>Touched: <span data-field="touched">{{ form.controls.myTextarea.touched }}</span></div>
-    <div>Dirty: <span data-field="dirty">{{ form.controls.myTextarea.dirty }}</span></div>
-    <div>Disabled: <span data-field="disabled">{{ form.controls.myTextarea.disabled }}</span></div>
-    <div>Value: <span data-field="value">{{ form.controls.myTextarea.value }}</span></div>
-    <div>Valid: <span data-field="valid">{{ form.controls.myTextarea.valid }}</span></div>
-    <div [@if]="submittedValue !== undefined">Submitted: <span data-field="submitted">{{ submittedValue }}</span></div>
   `,
   changeDetection: ChangeDetectionStrategy.OnPush,
   standalone: true,
@@ -36,23 +20,10 @@ import { PorscheDesignSystemModule } from '@porsche-design-system/components-ang
 })
 export class TextareaExampleFormComponent {
   form = new FormGroup({
-    myTextarea: new FormControl<string>('', { validators: Validators.required, nonNullable: true }),
+    myTextarea: new FormControl<string>(''),
   });
 
   submittedValue: any = undefined;
-
-  setValue(): void {
-    this.form.controls.myTextarea.setValue('a');
-  }
-
-  resetValue(): void {
-    this.form.controls.myTextarea.reset('');
-  }
-
-  toggleDisabled(): void {
-    const control = this.form.get('myTextarea')!;
-    control.disabled ? control.enable() : control.disable();
-  }
 
   onSubmit(): void {
     this.submittedValue = JSON.stringify(this.form.value);