+++ packages/components-vue/src/router/index.ts
@@ -123,6 +123,11 @@ export const routes: RouteType[] = [
     component: fromExamples.CheckboxExampleControlledModel,
   },
   {
+    path: '/checkbox-example-form',
+    name: 'Checkbox Example Form',
+    component: fromExamples.CheckboxExampleForm,
+  },
+  {
     path: '/flyout-basic',
     name: 'FlyoutBasic',
     component: fromPages.FlyoutBasicPage,
@@ -198,6 +203,11 @@ export const routes: RouteType[] = [
     component: fromExamples.InputPasswordExampleControlledModel,
   },
   {
+    path: '/input-password-example-form',
+    name: 'Input Password Example Form',
+    component: fromExamples.InputPasswordExampleForm,
+  },
+  {
     path: '/input-number-example',
     name: 'Input Number Example',
     component: fromExamples.InputNumberExample,
@@ -211,6 +221,11 @@ export const routes: RouteType[] = [
     path: '/input-number-example-controlled-model',
     name: 'Input Number Example Controlled Model',
     component: fromExamples.InputNumberExampleControlledModel,
+  },
+  {
+    path: '/input-number-example-form',
+    name: 'Input Number Example Form',
+    component: fromExamples.InputNumberExampleForm,
   },
   {
     path: '/input-date-example',
@@ -228,6 +243,11 @@ export const routes: RouteType[] = [
     component: fromExamples.InputDateExampleControlledModel,
   },
   {
+    path: '/input-date-example-form',
+    name: 'Input Date Example Form',
+    component: fromExamples.InputDateExampleForm,
+  },
+  {
     path: '/input-time-example',
     name: 'Input Time Example',
     component: fromExamples.InputTimeExample,
@@ -241,6 +261,11 @@ export const routes: RouteType[] = [
     path: '/input-time-example-controlled-model',
     name: 'Input Time Example Controlled Model',
     component: fromExamples.InputTimeExampleControlledModel,
+  },
+  {
+    path: '/input-time-example-form',
+    name: 'Input Time Example Form',
+    component: fromExamples.InputTimeExampleForm,
   },
   {
     path: '/input-text-example',
@@ -258,6 +283,11 @@ export const routes: RouteType[] = [
     component: fromExamples.InputTextExampleControlledModel,
   },
   {
+    path: '/input-text-example-form',
+    name: 'Input Text Example Form',
+    component: fromExamples.InputTextExampleForm,
+  },
+  {
     path: '/input-email-example',
     name: 'Input Email Example',
     component: fromExamples.InputEmailExample,
@@ -271,6 +301,11 @@ export const routes: RouteType[] = [
     path: '/input-email-example-controlled-model',
     name: 'Input Email Example Controlled Model',
     component: fromExamples.InputEmailExampleControlledModel,
+  },
+  {
+    path: '/input-email-example-form',
+    name: 'Input Email Example Form',
+    component: fromExamples.InputEmailExampleForm,
   },
   {
     path: '/input-tel-example',
@@ -288,6 +323,11 @@ export const routes: RouteType[] = [
     component: fromExamples.InputTelExampleControlledModel,
   },
   {
+    path: '/input-tel-example-form',
+    name: 'Input Tel Example Form',
+    component: fromExamples.InputTelExampleForm,
+  },
+  {
     path: '/input-url-example',
     name: 'Input Url Example',
     component: fromExamples.InputUrlExample,
@@ -303,6 +343,11 @@ export const routes: RouteType[] = [
     component: fromExamples.InputUrlExampleControlledModel,
   },
   {
+    path: '/input-url-example-form',
+    name: 'Input Url Example Form',
+    component: fromExamples.InputUrlExampleForm,
+  },
+  {
     path: '/input-search-example',
     name: 'Input Search Example',
     component: fromExamples.InputSearchExample,
@@ -316,6 +361,11 @@ export const routes: RouteType[] = [
     path: '/input-search-example-controlled-model',
     name: 'Input Search Example Controlled Model',
     component: fromExamples.InputSearchExampleControlledModel,
+  },
+  {
+    path: '/input-search-example-form',
+    name: 'Input Search Example Form',
+    component: fromExamples.InputSearchExampleForm,
   },
   {
     path: '/link-tile-example-hyphens',
@@ -373,6 +423,11 @@ export const routes: RouteType[] = [
     component: fromExamples.MultiSelectExampleDynamic,
   },
   {
+    path: '/multi-select-example-form',
+    name: 'Multi-Select Example Form',
+    component: fromExamples.MultiSelectExampleForm,
+  },
+  {
     path: '/pin-code-example',
     name: 'Pin Code Example',
     component: fromExamples.PinCodeExample,
@@ -386,6 +441,11 @@ export const routes: RouteType[] = [
     path: '/pin-code-example-controlled-model',
     name: 'Pin Code Example Controlled Model',
     component: fromExamples.PinCodeExampleControlledModel,
+  },
+  {
+    path: '/pin-code-example-form',
+    name: 'Pin Code Example Form',
+    component: fromExamples.PinCodeExampleForm,
   },
   {
     path: '/scroller-example',
@@ -408,6 +468,11 @@ export const routes: RouteType[] = [
     component: fromExamples.SegmentedControlExampleControlledModel,
   },
   {
+    path: '/segmented-control-example-form',
+    name: 'Segmented Control Example Form',
+    component: fromExamples.SegmentedControlExampleForm,
+  },
+  {
     path: '/radio-group-example',
     name: 'Radio Group Example',
     component: fromExamples.RadioGroupExample,
@@ -421,6 +486,11 @@ export const routes: RouteType[] = [
     path: '/radio-group-example-controlled-model',
     name: 'Radio Group Example Controlled Model',
     component: fromExamples.RadioGroupExampleControlledModel,
+  },
+  {
+    path: '/radio-group-example-form',
+    name: 'Radio Group Example Form',
+    component: fromExamples.RadioGroupExampleForm,
   },
   {
     path: '/select-example',
@@ -446,6 +516,11 @@ export const routes: RouteType[] = [
     path: '/select-example-required',
     name: 'Select Example Required',
     component: fromExamples.SelectExampleRequired,
+  },
+  {
+    path: '/select-example-form',
+    name: 'Select Example Form',
+    component: fromExamples.SelectExampleForm,
   },
   {
     path: '/select-example-async-load',
@@ -501,6 +576,11 @@ export const routes: RouteType[] = [
     path: '/textarea-example-controlled-model',
     name: 'Textarea Example Controlled Model',
     component: fromExamples.TextareaExampleControlledModel,
+  },
+  {
+    path: '/textarea-example-form',
+    name: 'Textarea Example Form',
+    component: fromExamples.TextareaExampleForm,
   },
   {
     path: '/text-field-wrapper-example-imask',