+++ packages/components/src/components/input-number/input-number.tsx
@@ -17,6 +17,7 @@ import {
   FORM_STATES,
   getPrefixedTagNames,
   hasPropValueChanged,
+  implicitSubmit,
   THEMES,
   validateProps,
 } from '../../utils';
@@ -217,6 +218,7 @@ export class InputNumber {
         onInput={this.onInput}
         onChange={this.onChange}
         onBlur={this.onBlur}
+        onKeyDown={this.onKeyDown}
         name={this.name}
         form={this.form}
         type="number"
@@ -267,6 +269,10 @@ export class InputNumber {
       />
     );
   }
+
+  private onKeyDown = (e: KeyboardEvent): void => {
+    implicitSubmit(e, this.internals, this.host);
+  };
 
   private onChange = (e: Event): void => {
     e.stopPropagation();