+++ packages/storefront/tests/unit/specs/url.spec.ts
@@ -29,7 +29,8 @@ describe('url', () => {
           expect(url.startsWith('https://www.')).toBe(false);
         });
 
-        test('should be reachable', async () => {
+        // TODO: Remove skipIf once website is loading normally again
+        test.skipIf(url.includes('https://adrianroselli.com/'))('should be reachable', async () => {
           const { status } = await fetch(url);
 
           expect(status).not.toBe(404);